refactor(playback): organize backends into folders
